requirements.txt
asgiref==3.9.1
catboost==1.2.8
charset-normalizer==3.4.3
contourpy==1.3.2
cycler==0.12.1
Django==5.2.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
fonttools==4.59.1
graphviz==0.21
joblib==1.5.1
kiwisolver==1.4.9
lightgbm==4.6.0
lxml==6.0.0
matplotlib==3.10.5
narwhals==2.1.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
reportlab==4.4.3
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.14.1
tzdata==2025.2
xgboost==3.0.4
